Шифрование на цпу и не только

+++ app/core/encryption/math_utils.hpp
@@ -15,7 +15,7 @@ using uint32 = uint32_t;
 using uint16 = uint16_t;
 using uint8 = uint8_t;
 using byte = unsigned char;
-using word = unsigned long long;
+using word = unsigned int;
 
 inline auto operator ""_GB(size_t const x) { return 1024 * 1024 * 1024 * x; }
